migraciones locas, controladores locos y algún producto

+++ Notify/Models/Producto.cs
@@ -12,8 +12,8 @@ namespace Notify.Models
         public int codigo { get; set; } 
         [StringLength(50)]
         public String nombre { get; set; }
-        public double precio { get; set; }
-
+       
+        public decimal precio { get; set; }
 
         public virtual List<Linea> envio_nombre_producto { get; set; } = new List<Linea>();
     }
